requirements.txt
line-bot-sdk==3.7
flask==3.0.0
psycopg2-binary
